fix: ci process (#2)
**Issue #, if available:**

## Description of changes:

The CI process is currently broken due to a missing region parameter. On
the local machine it defaulted back to the "configured" region and
therefor it did not fail locally.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.