fix: environment sorting issue (#19)
**Issue #, if available:**

## Description of changes:

There was a small bug in the landingzone-organization module, that has
been fixed in 0.6.3. Also optimized the code to use the workload class
from the landingzone-organization module. And added some documentation
on how to use the ordering functionality.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.