fix: use AWS_REGION environment variable (#36)
**Issue #, if available:**

## Description of changes:

When no named profile is used we will need to provide a region for the
session. By falling back on the `AWS_REGION` environment variable we can
easily support a proper flow.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.